+++ packages/saltcorn-markup/table.ts
@@ -4,8 +4,22 @@
  */
 
 import tags = require("./tags");
-const { a, td, tr, th, text, div, table, thead, tbody, ul, li, span, h4 } =
-  tags;
+const {
+  a,
+  td,
+  tr,
+  th,
+  text,
+  div,
+  table,
+  thead,
+  tbody,
+  ul,
+  li,
+  span,
+  h4,
+  style,
+} = tags;
 import helpers = require("./helpers");
 import type { SearchBarOpts, RadioGroupOpts } from "./helpers";
 const { pagination } = helpers;
@@ -17,14 +31,15 @@ const { pagination } = helpers;
 const headerCell = (hdr: any): string =>
   th(
     (hdr.align || hdr.width) && {
-      style:
-        (hdr.align ? `text-align: ${hdr.align};` : "") +
-        (hdr.width ? `width: ` + hdr.width : ""),
+      style: hdr.width ? `width: ` + hdr.width : "",
+      ...(hdr.align ? { class: `text-align-${hdr.align}` } : {}),
     },
     hdr.sortlink
       ? span({ onclick: hdr.sortlink, class: "link-style" }, hdr.label)
       : hdr.label
   );
+
+const headerFilter = (hdr: any): string => th(hdr.header_filter || null);
 
 // declaration merging
 namespace TableExports {
@@ -33,6 +48,7 @@ namespace TableExports {
     key: string | Function;
     width?: string;
     align?: string;
+    header_filter?: string;
   };
 
   export type OptsParams = {
@@ -47,6 +63,9 @@ namespace TableExports {
     tableClass?: string;
     tableId?: string;
     grouped?: string;
+    header_filters?: boolean;
+    responsiveCollapse?: boolean;
+    collapse_breakpoint_px?: number;
   };
 }
 type HeadersParams = TableExports.HeadersParams;
@@ -107,9 +126,9 @@ const mkTable = (
         td(
           {
             style: {
-              ...(hdr.align ? { "text-align": hdr.align } : {}),
               ...(hdr.width && opts.noHeader ? { width: hdr.width } : {}),
             },
+            ...(hdr.align ? { class: `text-align-${hdr.align}` } : {}),
           },
           typeof hdr.key === "string" ? text(v[hdr.key]) : hdr.key(v)
         )
@@ -139,16 +158,69 @@ const mkTable = (
       },
       !opts.noHeader &&
         !opts.transpose &&
-        thead(tr(hdrs.map((hdr: HeadersParams) => headerCell(hdr)))),
+        thead(
+          tr(hdrs.map((hdr: HeadersParams) => headerCell(hdr))),
+          opts.header_filters
+            ? tr(
+                { class: "header-filters" },
+                hdrs.map((hdr: HeadersParams) => headerFilter(hdr))
+              )
+            : null
+        ),
       tbody(
         opts.transpose
           ? transposedBody(hdrs, vs, opts)
           : opts.grouped
-          ? groupedBody(vs)
-          : (vs || []).map(val_row)
+            ? groupedBody(vs)
+            : (vs || []).map(val_row)
       )
     ),
-    opts.pagination && pagination(opts.pagination)
+    opts.pagination && pagination(opts.pagination),
+    //https://css-tricks.com/responsive-data-tables/
+    opts.responsiveCollapse &&
+      opts.tableId &&
+      style(`@media 
+only screen and (max-width: ${opts.collapse_breakpoint_px || 760}px) {
+	#${opts.tableId} table, #${opts.tableId} thead, #${opts.tableId} tbody, #${opts.tableId} th, #${opts.tableId} td, #${opts.tableId} tr { 
+		display: block; 
+	}
+
+  #${opts.tableId} tr.header-filter {
+    display: none;
+  }
+  #${opts.tableId} td.text-align-right,
+  #${opts.tableId} td.text-align-right,
+  #${opts.tableId} th.text-align-center,
+  #${opts.tableId} th.text-align-center {
+     text-align: left !important;
+  }
+
+	#${opts.tableId} thead tr { 
+		position: absolute;
+		top: -9999px;
+		left: -9999px;
+	}
+	
+	#${opts.tableId} tr { border: 1px solid #ccc; }
+	
+	#${opts.tableId} td { 
+		border: none;
+		border-bottom: 1px solid #eee; 
+		position: relative;
+		padding-left: 50%; 
+	}
+	
+	#${opts.tableId} td:before { 
+		position: absolute;
+		top: 6px;
+		left: 6px;
+		width: 45%; 
+		padding-right: 10px; 
+		white-space: nowrap;
+	}
+
+  ${hdrs.map((hdr: HeadersParams, ix: number) => `#${opts.tableId} td:nth-of-type(${ix + 1}):before { content: "${hdr.label}"; }`).join("\n")}	
+}`)
   );
 };